% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.r
\name{col_align}
\alias{col_align}
\title{Align an ANSI colored string}
\usage{
col_align(text, width = getOption("width"), align = c("left", "center",
  "right"))
}
\arguments{
\item{text}{The character vector to align.}

\item{width}{Width of the field to align in.}

\item{align}{Whether to align `"left"`, `"center"` or `"right"`.}
}
\value{
The aligned character vector.
}
\description{
Align an ANSI colored string
}
\examples{
col_align(red("foobar"), 20, "left")
col_align(red("foobar"), 20, "center")
col_align(red("foobar"), 20, "right")
}
