% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.r
\name{col_strwrap}
\alias{col_strwrap}
\title{Helper function for wrapping crayon text}
\usage{
col_strwrap(x, width = 0.9 * getOption("width"), indent = 0, exdent = 0,
  prefix = "", simplify = TRUE, initial = prefix)
}
\arguments{
\item{x}{a character vector, or an object which can be converted to a
    character vector by \code{\link{as.character}}.}

\item{width}{a positive integer giving the target column for wrapping
    lines in the output.}

\item{indent}{a non-negative integer giving the indentation of the
    first line in a paragraph.}

\item{exdent}{a non-negative integer specifying the indentation of
    subsequent lines in paragraphs.}

\item{prefix}{a character string to be used as prefix for
    each line except the first, for which \code{initial} is used.}

\item{simplify}{a logical.  If \code{TRUE}, the result is a single
    character vector of line text; otherwise, it is a list of the same
    length as \code{x} the elements of which are character vectors of
    line text obtained from the corresponding element of \code{x}.
    (Hence, the result in the former case is obtained by unlisting that
    of the latter.)}

\item{initial}{a character string to be used as prefix for
    each line except the first, for which \code{initial} is used.}
}
\value{
a character vector of line content after wrapping
}
\description{
A drop-in replacement for \code{strwrap} for use with stylized crayon text.
}
\examples{

text <- paste(rep(c(red("red"), yellow("yellow"), green("green"), 
    cyan("cyan"), blue("blue"), magenta("magenta")), 5), collapse = " ")

wrapped_crayon_output <- col_strwrap(
  text, 
  indent = 1, 
  exdent = 3, 
  initial = bgBlack(white("<#> ")),
  prefix  = bgBlack(white(" #> ")))

cat(paste(wrapped_crayon_output, collapse = "\\n"))

}
\seealso{
Other ANSI string operations: \code{\link{col_align}},
  \code{\link{col_nchar}}, \code{\link{col_strsplit}},
  \code{\link{col_substring}}, \code{\link{col_substr}}
}
