% Generated by roxygen2 (4.0.1): do not edit by hand
\name{concat}
\alias{\%+\%}
\alias{concat}
\title{Concatenate strings}
\usage{
lhs \%+\% rhs
}
\arguments{
\item{lhs}{Left hand side, character scalar.}

\item{rhs}{Right hand side, character scalar.}
}
\value{
Concatenated string.
}
\description{
Concatenate strings
}

