% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{crayon}
\alias{bgBlack}
\alias{bgBlue}
\alias{bgCyan}
\alias{bgGreen}
\alias{bgMagenta}
\alias{bgRed}
\alias{bgWhite}
\alias{bgYellow}
\alias{black}
\alias{blue}
\alias{bold}
\alias{crayon}
\alias{crayon-package}
\alias{cyan}
\alias{dark}
\alias{green}
\alias{hidden}
\alias{inverse}
\alias{italic}
\alias{magenta}
\alias{red}
\alias{reset}
\alias{silver}
\alias{strikethrough}
\alias{underline}
\alias{white}
\alias{yellow}
\title{Colored terminal output}
\arguments{
\item{...}{Strings to style.}
}
\description{
TODO: document 'crayon.enabled' option
}
\details{
Styles: reset bold dark italic underline inverse hidden strikethrough
black red green yellow blue magenta cyan white silver
bgBlack bgRed bgGreen bgYellow bgBlue bgMagenta bgCyan bgWhite.
}

