% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinery.r
\name{drop_style}
\alias{drop_style}
\title{Remove a style}
\usage{
drop_style(style)
}
\arguments{
\item{style}{The name of the style to remove. No error is given
for non-existing names.}
}
\value{
Nothing.
}
\description{
Remove a style
}
\examples{
make_style(new_style = "maroon", bg = TRUE)
cat(style("I am maroon", "new_style"), "\n")
drop_style("new_style")
"new_style" \%in\% names(styles())
}
\seealso{
Other styles: 
\code{\link{make_style}()}
}
\concept{styles}
