% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_ansi.r
\name{has_style}
\alias{has_style}
\title{Check if a string has some ANSI styling}
\usage{
has_style(string)
}
\arguments{
\item{string}{The string to check. It can also be a character
vector.}
}
\value{
Logical vector, \code{TRUE} for the strings that have some
ANSI styling.
}
\description{
Check if a string has some ANSI styling
}
\examples{
## The second one has style if crayon is enabled
has_style("foobar")
has_style(red("foobar"))
}
