% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{hyperlink}
\alias{hyperlink}
\alias{has_hyperlink}
\title{Terminal Hyperlinks}
\usage{
hyperlink(text, url)

has_hyperlink()
}
\arguments{
\item{text}{Text to show. \code{text} and \code{url} are recycled to match their
length, via a \code{paste0()} call.}

\item{url}{URL to link to.}
}
\value{
Logical scalar, for \code{has_hyperlink()}.
}
\description{
Terminal Hyperlinks
}
\details{
hyperlink()` creates an ANSI hyperlink.

\code{has_hyperlink()} checks if the current \code{stdout()} supports hyperlinks.
terminal links.

See also
\url{https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda}.
}
\examples{
cat("This is an", hyperlink("R", "https://r-project.org"), "link.\n")
has_hyperlink()
}
