% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show_ansi_colors}
\alias{show_ansi_colors}
\title{Show the ANSI color table on the screen}
\usage{
show_ansi_colors(colors = num_colors())
}
\arguments{
\item{colors}{Number of colors to show, meaningful values
are 8 and 256. It is automatically set to the number of
supported colors, if not specified.}
}
\value{
The printed string, invisibly.
}
\description{
Show the ANSI color table on the screen
}
