% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts.r
\name{start}
\alias{finish}
\alias{start}
\title{Switch on or off a style}
\usage{
start(x, ...)

finish(x, ...)
}
\arguments{
\item{x}{Style.}

\item{...}{Ignored.}
}
\description{
Make a style active. The text printed to the screen from now
on will use this style.
}
\details{
This function is very rarely needed, e.g. for colored user
input. For other reasons, just call the style as a function on
the string.
}
\examples{
## The input is red (if color is supported)
get_name <- function() {
  cat("Enter your name:", start(red))
  input <- readline()
  cat(finish(red))
  input
}
name <- get_name()
name
}

