% Generated by roxygen2 (4.0.1): do not edit by hand
\name{style}
\alias{style}
\title{Add style to a string}
\usage{
style(string, as)
}
\arguments{
\item{string}{Character vector to style.}

\item{as}{Style function to apply, either the function object,
or its name.}
}
\value{
Styled character vector.
}
\description{
See \code{names(styles)}, or the crayon manual for available styles.
}
\examples{
## These are equivalent
style("foobar", bold)
style("foobar", "bold")
bold("foobar")
}

