% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{chr}
\alias{chr}
\title{Convert to character}
\usage{
chr(x, ...)
}
\arguments{
\item{x}{Object to be coerced.}

\item{...}{Further arguments to pass to \code{as.character()}.}
}
\value{
Character value.
}
\description{
This function just calls \code{\link[=as.character]{as.character()}}, but it is
easier to type and read.
}
