% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{concat}
\alias{concat}
\alias{\%+\%}
\title{Concatenate character vectors}
\usage{
lhs \%+\% rhs
}
\arguments{
\item{lhs}{Left hand side, character vector.}

\item{rhs}{Right hand side, character vector.}
}
\value{
Concatenated vectors.
}
\description{
The length of the two arguments must match, or
one of them must be of length one. If the length of
one argument is one, then the output's length will
match the length of the other argument. See examples
below.
}
\examples{
"foo" \%+\% "bar"

letters[1:10] \%+\% chr(1:10)

letters[1:10] \%+\% "-" \%+\% chr(1:10)

## This is empty (unlike for parse)
character() \%+\% "*"
}
