% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_color.r
\name{has_color}
\alias{has_color}
\title{Does the current R session support ANSI colors?}
\usage{
has_color()
}
\value{
\code{TRUE} if the current R session supports color.
}
\description{
From crayon 2.0.0, this function is simply a wrapper on
\code{\link[=num_ansi_colors]{num_ansi_colors()}}.
}
\examples{
has_color()
}
