% Generated by roxygen2 (4.0.2): do not edit by hand
\name{num_colors}
\alias{num_colors}
\title{Number of colors the terminal supports}
\usage{
num_colors()
}
\value{
Numeric scalar, the number of colors the terminal supports.
}
\description{
We use the \code{tput} shell command to detect the
number of colors. If \code{tput} is not available,
but we think that the terminal supports colors, then
eigth colors are assumed.
}

