% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_color.r
\name{num_colors}
\alias{num_colors}
\title{Number of colors the terminal supports}
\usage{
num_colors(forget = FALSE)
}
\arguments{
\item{forget}{Whether to forget the cached result of the color check.}
}
\value{
Numeric scalar, the number of colors the terminal supports.
}
\description{
Number of colors the terminal supports
}
\details{
If the \code{crayon.colors} option is set, then we
just use that. It should be an integer number.

In Emacs, we report eight colors.

Otherwise, we use the \code{tput} shell command to detect the
number of colors. If \code{tput} is not available,
but we think that the terminal supports colors, then
eigth colors are assumed.

For efficiency, \code{num_colors} caches its result. To
re-check the number of colors, set the \code{forget} argument to
\code{TRUE}.
}
\examples{
num_colors()
}

