% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_color.r
\name{num_colors}
\alias{num_colors}
\title{Number of colors the terminal supports}
\usage{
num_colors(forget = FALSE)
}
\arguments{
\item{forget}{Ignored. Included for backwards compatibility.}
}
\value{
Number of ANSI colors.
}
\description{
From crayon version 2.0.0, this function is a simple wrapper on
\code{\link[=num_ansi_colors]{num_ansi_colors()}}, with the additional twist that the \code{crayon.colors}
option is still obseved, and takes precedence, for compatibility.
}
\examples{
num_colors()
}
