% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{styles}
\alias{styles}
\title{ANSI escape sequences of crayon styles}
\format{A named list. Each list element is a list of two
  strings, named \sQuote{open} and \sQuote{close}.}
\usage{
styles
}
\description{
You can use this object to list all availables crayon styles,
via \code{names(styles)}, or to explicitly apply an ANSI
escape seauence to a string.
}
\examples{
names(styles)
cat(styles[["bold"]]$close)
}
\seealso{
\code{\link{crayon}} for the beginning of the crayon manual.
}
\keyword{datasets}

