% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinery.r
\name{styles}
\alias{styles}
\title{ANSI escape sequences of crayon styles}
\usage{
styles()
}
\value{
A named list. Each list element is a list of two
strings, named \sQuote{open} and \sQuote{close}.
}
\description{
You can use this function to list all availables crayon styles,
via \code{names(styles())}, or to explicitly apply an ANSI
escape seauence to a string.
}
\examples{
names(styles())
cat(styles()[["bold"]]$close)
}
\seealso{
\code{\link[=crayon]{crayon()}} for the beginning of the crayon manual.
}
